requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
plotly==5.18.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
psycopg2-binary==2.9.9
alembic==1.12.1
